+++ src/py/bbctrl/PlanTimer.py
@@ -35,19 +35,18 @@ log = logging.getLogger('PlanTimer')
 class PlanTimer(object):
     def __init__(self, ctrl):
         self.ctrl = ctrl
+        self.plan_times = None
 
         self.reset()
-        self._report()
-
         self.ctrl.state.set('plan_time', 0)
         ctrl.state.add_listener(self._update)
+        self._report()
 
 
     def reset(self):
         self.plan_time = 0
         self.move_start = None
         self.hold_start = None
-        self.plan_times = None
         self.plan_index = 0
 
 
@@ -66,42 +65,53 @@ class PlanTimer(object):
 
                 self.ctrl.state.set('plan_time', round(t))
 
-        self.timer = self.ctrl.ioloop.call_later(1, self._report)
+        self.ctrl.ioloop.call_later(1, self._report)
+
+
+    def _update_state(self, state):
+        if state in ['READY', 'ESTOPPED']:
+            self.ctrl.state.set('plan_time', 0)
+            self.reset()
+
+        elif state == 'HOLDING': self.hold_start = time.time()
+        elif (state == 'RUNNING' and self.hold_start is not None and
+              self.move_start is not None):
+            self.move_start += time.time() - self.hold_start
+            self.hold_start = None
+
+
+    def _update_times(self, filename):
+        if not filename: return
+        future = self.ctrl.preplanner.get_plan(filename)
+
+        def cb(future):
+            if (filename != self.ctrl.state.get('selected') or
+                future.cancelled()): return
+
+            self.reset()
+            path, meta = future.result()
+            self.plan_times = meta['times']
+
+        self.ctrl.ioloop.add_future(future, cb)
+
+
+    def _update_time(self, currentID):
+        if self.plan_times is None: return
+
+        while self.plan_index < len(self.plan_times):
+            id, t = self.plan_times[self.plan_index]
+            if id <= currentID: self.move_start = time.time()
+            if currentID <= id: break
+            self.plan_time = t
+            self.plan_index += 1
 
 
     def _update(self, update):
         # Check state
-        if 'xx' in update:
-            state = update['xx']
-
-            if state in ['READY', 'ESTOPPED']:
-                self.ctrl.state.set('plan_time', 0)
-                self.reset()
-
-            elif state == 'HOLDING': self.hold_start = time.time()
-            elif (state == 'RUNNING' and self.hold_start is not None and
-                  self.move_start is not None):
-                self.move_start += time.time() - self.hold_start
-                self.hold_start = None
+        if 'xx' in update: self._update_state(update['xx'])
 
         # Get plan times
-        if self.plan_times is None or 'selected' in update:
-            active_plan = self.ctrl.state.get('selected', '')
-
-            if active_plan:
-                plan = self.ctrl.preplanner.get_plan(active_plan)
-
-                if plan is not None and plan.done():
-                    self.reset()
-                    self.plan_times = plan.result()[1]
+        if 'selected' in update: self._update_times(update['selected'])
 
         # Get plan time for current id
-        if self.plan_times is not None and 'id' in update:
-            currentID = update['id']
-
-            while self.plan_index < len(self.plan_times):
-                id, t = self.plan_times[self.plan_index]
-                if id <= currentID: self.move_start = time.time()
-                if currentID <= id: break
-                self.plan_time = t
-                self.plan_index += 1
+        if 'id' in update: self._update_time(update['id'])